Return Seq, not Vec, from PriorityEncoderOH
It doesn't really make sense to create a one-hot result then follow it
with a decoder.  It's more performant to use a PriorityEncoder followed
by a comparator.  Discourage the former by returning Seq, not Vec.

In Chisel2, it seems the return type was originally Seq, but was at some
point inadvertently changed to Vec.