Sketch Implementation of SourceInfo
First, this introduces the SourceInfo case class which contains an
Option[SourceInfo.SourceLoc]. The only current SourceInfo.SourceLoc is
FromFile(file: String, line: String).

A SourceInfo will be implicitly materialized via a macro if one is not
found earlier during implicit resolution. (An import or enclosing
implicit SourceInfo would be earlier.)

This change specifically adds implicit SourceInfo to Wire as a
demonstration. Wire was deemed 'safe' because you are extraordinarily
unlikely to 'apply' directly after a Wire. Notice how all the internal
uses of Wire had to have SourceInfo more directly supplied since
materialize could not be used due to macro resolution rules. (The
explicit SourceInfo suppression should be replaced by implicits in the
function calling wiring; however, some are liable to be followed by
'apply' and thus it is unsafe to do so without backwards compatibility
concerns)

PS: Ideally, the implicit materialize function would be 'hidden'
internally. This is partly why it is advised to import a subpackage of
Chisel and not just Chisel. Am contemplating introducing an intentional
implicit resolution conflict in the internal package object as a
stopgap.