Fixup to Chisel connections, literal creation, direction, etc.
First set of changes:
Require io be wrapped in IO, add Binding to track how/if Data is bound

* Adds internal.Binding, which is a marker on Elements (subtype/leaves of
Data), that tracks whether a Data has been placed onto the hardware graph
as a register, port, literal, wire, operation result, etc.
* This requires the io of a module actually be explicitly placed on the graph
AS it is being created. (Since the module will want to connect to its own IO,
module close is too late to do the binding.)
* The binding check will also enforce that a bundle constructor does not
actually create logic that is interpreted as part of the 'core bundle'.
* Subsequent CLs will expand upon the Binding system to carry more data
and execute more checks.
* Now, frontend emissions of undefined declarations should be classed as bugs
that should be caught by this system.

Second set of changes:
First, a listing of API changes:
* For Literals,
-- You can no longer create a literal with UInt|SInt|Bool.apply(...)
++ Literal UInts/SInts are created from scala Int/BigInt by e.g. 0.asUInt,
myInt.asUInt(5). The latter means forced width to be 5.
++ Literal UInt can also be created from String bu asUInt
++ Literal Bools are created from scala Boolean by e.g. true.asBool
++ The UInt|SInt|Bool factory objects have Lit methods (which is what the old
literal apply methods were renamed to. These are primarily intended for
internal and transitional use. The .as* functions should be preferred.
++ The U|S|B methods added to Int|BigInt|String (identical to asUInt, asSInt,
asBool) have been added and marked as deprecated.

* For working with Chisel Data,
-- You can no longer supply a direction to UInt|SInt|Bool
-- You can no longer use .asInput|.asOutput|.flip. They have been marked as
deprecated.
++ Instead, wrap in Input(), Output(), Flipped() call. These functions can only
be called on unbound data (avoiding extranneous copies)
++ In conjunction with literal changes, UInt(#), SInt(#) now always creates an
unbound variant of that width ie. equivalent to UInt(width=#), etc.

* Particularly special notice on cloneType and newType,
-- cloneType is now protected and should not be called by the user
-- The user must still define cloneType (usually) for their cloned Bundles
however the contract has simiplified in that the clone merely must be
'structurally' similar to this.
++ Calling newType will invoke cloneType to get the structural copy and then
return that copy with the directions cleaned up for you.
++ This cleans up a somewhat nasty bug where, for example, direction changes
could get lost when a Vec of a directioned Data was created.
++ Reg and Wire now requires that outType/t variable be unbound (to avoid
potential user errors). newType is still called on it so the same Type can be
used to create multiple registers and wires.

* On connection semantics and directionality,
++ Internal nodes like Reg and Wire that have no defined notion of input or
output direction now lose any direction information when going from Unbound to
Bound. It may be enforced later that the types be explicitly undirectioned.
++ <> is now back to being a bidirectional connect mean to operate between
Module IOs only
++ <> requires Bundles on both sides to have exactly the same fields
++ := is now back to being a monodirectional connect that assumes all
subelements of the RHS are sources for corresponding subelement sinks on the
LHS.
++ := requires the source Bundle to have at least the fields of the sink
Bundle.

* In ChiselUtils,
++ EnqIO and DeqIO have been reworked and are no longer distinct classes and
just factory methods for creating an appropriate DecoupledIO
++ Decoupled factory object renamed to DecoupledIO
++ DecoupledIO fire() method renamed to firing
++ DecoupledIO has enq, noenq, deq, nodeq, similar to old EnqIO, DeqIO methods
++ ValidIO type renamed to Valid, no longer forces Output

Second, some notes on implementation changes.
Direction is no longer stored on the Data but on the Binding attached to the
data. The concept of Binding has been enriched to also introduce a 'Binder'
function that a caller users to describe how they want a Data's nodes to be
bound dependent on the specifics of the UnboundBinding. For example, PortBinder
will create a Binding with the same direction information of the
UnboundBinding.

The MonoConnect and BiConnect operators will only emit FIRRTL connections
for Elements, never Aggregates. This is to avoid the issues with FIRRTL
connection semantics. Relatedly, Bundles are never marked as flip or input.
Only Elements of Bundles (or Vec chains of Elements in Bundles) can be marked
as flip. Essentially, all flips have no been distributed through the Bundles to
the root nodes.

Vec types now contain a sample_element that tracks all the binding changes to
the actual elements. This sample_element is used when emitting the type of the
Vec to FIRRTL and when creating a dynamically indexed port. This is needed to
handle the case of an empty Vec having no elements but still requiring those
descriptions. For consistency, this sample_element is used even when the Vec is
not empty.

cloneTypeWidth behavior is somewhat weird. However, this is private and only
used to create operators and should not be an immediate issue. This will
eventually be removed (since cloneTypeWidth previously and currently will not
create appropriate Muxes of Aggregates).

The use of DynamicVariable in Builder has been more carefully hidden from the
rest of the internals by making dynamicContext private and adding accessors to
the other global state.

I want to watch the overhead in cloneType and newType to make sure too much
time isn't being spent making copies, manipulating them, copying again, as so
on as people built more complex types. (Not having Input, Output, Flipped
create a copy was inspired by a desire to avoid these copies)