Remove unused BuildInfo.BuildTime.
BuildInfo.BuildTime appears to cause sbt to always recompile BuildInfo to get
the latest time, which causes a recompile on any package that depends on
chisel3. This is quite painful for doing development on any Chisel projects.
BuildTime doesn't actually appear to be in use, and removing it fixes the
unnecessary recompilation issue.