Immutable parsing, add file disable output args
This switches options parsing in ChiselExecutionOptions and Driver to use
new immutable parsing provided by FIRRTL. This aims to provide a single
API based on command line options for interacting with Chisel/FIRRTL and
not exposing internal case class members via an options manager.

This adds two new options for controlling whether or not specific output
files are generated:
  --dont-save-chirrtl
  --dont-save-annotations

This deprecates passing in an explicit options manager to Driver.execute.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>