Un-deprecate OptionsManager Driver.execute
The removes the deprecation on Driver.execute(scopt.OptionsManager) and
uses this as the backend for Driver.execute(Array[String]).

This is enabled by options checks in FirrtlExecutionOptions that operate
only on annotations and not on command line options. Since this then
enables multi-stage option construction, the former Driver.execute no
longer requires deprecation.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>