Add Chisel circuit to ChiselExecutionOptions
This adds the Chisel-generated Circuit (not CHIRRTL) to
ChiselExecutionOptions and includes an associated annotation for this. Due
to the inability of this annotation to be serialized, it is not exposed
externally as a CHIRRTL annotation.

This adds tests for this annotation passed via an explicit annotation or
the "--dut" command line option.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>