Add Chisel3 assembly (fat jar) to sbt build
This adds (the option) to build a fat jar via sbt-assembly. This will be
built into utils/bin/chisel3.jar.

This enables full command-line usage of Chisel to build circuits that take
no parameters, i.e., it allows a user to use the `--dut` option. The
following example illustrates building one of the test modules directly
from the command line:

  java -cp ./utils/bin/chisel3.jar:./target/scala-2.11/test-classes \
    chisel3.Driver \
    --dut chiselTests.DummyModule

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>