Add SourceInfoDoc trait w/ ScalaDoc group
This adds a bare trait that provides a ScalaDoc group,
`SourceInfoTransformMacro`, that is intended to group all `do_*` methods
in one place in the documentation. This provides a group description for
this group and a priority (1001) which is lower than all other normal
groups (the lowest priority is supposed to be 1000).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>