ChiselDutGeneratorAnnotation, elab/convert in Anno
This adds a ChiselDutGeneratorAnnotation that can be used to store an
anonymous function that returns a RawModule. Due to the inability of this
annotation to be serialized, it extends the Unserializable FIRRTL trait.
In addition to the mandatory toJsonSerializable method, this provides
elaborate and convert methods for elaborating a module to a Chisel circuit
and converting it to a FIRRTL circuit.

The existing ChiselCircuitAnnotation is updated to be Unserializable and
provide a convert method to emit a FIRRTL circuit.

Chisel's Driver is reduced in complexity by using the methods of these
annotations for elaboration and conversion.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>