Better error message on missing BlackBox resource
This wraps interactions with a BlackBox resource file such that a
FileNotFoundException are wrapped in a BlackBoxNotFoundException and
rethrown. This provides a better, verbose error message to the user and avoids a
FileNotFoundException showing up as an internal FIRRTL error.

This adds tests that the expected exception is thrown for both
BlackBoxResourceAnno and BlackBoxResourceAnno.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>