Better error message for UninferredWidth exception
This changes the CheckTypes.UniferredWidth exception to include the pretty
printed Target that was uninferred and suggests to the user that they may
have forgotten to assign to it. This changes the CheckTypes pass to
communicate the necessary Target information during AST traversal such
that when an uninferred width is found, the Target is known and available.

This also adds one test checking the message of the UniferredWidth
exception.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>