Fix LoadMemoryTransform for Instance Annotations (#914)
Quickfix for LoadMemoryTransform that gets this to work with Instance
Annotations. The new Instance Annotations caused a corner case where a
LoadMemoryAnnotation would be duplicated (via update/renaming) and the
resulting annotation would differ from the original in only their
originalMemoryNameOpt field. This corrects that by having the
ChiselLoadMemoryAnnotation also emit the originalMemoryNameOpt field where
it did not previously.

First part of a fix for freechipsproject/firrtl#922.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>