Add prettyPrint method to Target
This adds a pretty printer for firrtl.annotation.Target and associated
tests. This uses a tree-like output where the following target
~Circuit|Module/foo:Foo>ref.field[0] will serialize to:

  circuit Circuit:
  └── module Module:
      └── foo of Foo:
          └── ref.field[0]

This enables better error messages and a human readable syntax better than
the existing serialize method (and avoiding the need for users to
understand the Target serialization syntax), but that is not intended to
be deserialized nor space efficient.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>