Add factory functions for Valid values to RegEnable
The Valid type is often passed to modules, and is an obvious choice
for setting a register based on a value and an enable signal. These
factory functions simply act as a boilerplate shorthand for
construction registers in a less verbose way.

Both functions provided simply map the Valid to either of the two
existing RegEnable functions.