Fork all sbt test and run tasks (#1002)
This avoids an apparent problem somewhere in sbt and/or scalatest
where the JVM runs out of metaspace if you repeatedly run tasks. This
is an annoyance for FIRRTL developers or users that keep an sbt
session open. This kludges around that by forking all tasks into a
separate JVM.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>