Do Shr constant propagation in Legalize
This uses the foldShiftRight method of the ConstantPropagation
Transform when legalizing Shr PrimOps. This has the effect of removing
literals with bit extracts from the MinimumVerilogCompiler.

This makes the formerly private foldShiftRight method of a public
method of the ConstantPropagation companion object.

Tests in the MimimumVerilogCompilerSpec are updated to check that Shr
is handled as intended.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>