Fix BoringUtils.bore for internal boring
This fixes a bug where internal boring using BoringUtils.bore would
fail because it was using instanceName which cannot be called before
the module closes. Previously, this meant that BoringUtils.bore would
work for boring instances (which are closed in a parent), but not for
boring signals in the current, unclosed module.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>