Use per-project Scala version, explicit tags
This changes the build process to use whatever the default Scala
version is as defined by the specific project/tag being built. This
contrasts with using an unsafe, explicit '++ 2.12.6' which may fail if
dependencies for that Scala version aren't available.

This changes documentation to build using explicit tags specific in
the Makefile (for old versions and SNAPSHOT). The "latest" version is
then defined as a symlink against an already built version as opposed
to the version in the submodule.

In effect, this removes submodules from contributing to the Scaladoc
build process.