Record self-renames in RenameMap, distinct renames
Change the behavior of RenameMap.completeTarget so that self-renames
do not silently *not* happen. Previously, requests to self-rename
would be ignored unless they were packaged in a sequences of renames
that included a self-rename.

Change renames to be recorded distinctly so that multiple requests to
rename to the same thing will now deduplicate. Previously, these
renames would be recorded multiple times. This change was required
because allowing self-renames exposed a bug in InferWidthsAnnosSpec
due to multiple renames.

These changes benefit the situation where you rightly want to do a
self-rename. Namely, when doing module duplication (with the
EliminateTargetPaths transform).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

Do not record the same rename multiple times

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>