Check EmitAnnotation class before emitting
Fixes a bug where an Emitter was only checking for the presence of an
EmitCircuitAnnotation or EmitAllModulesAnnotation to control its
emission flavor (one-file-per-module or one-file). This changes the
check to ensure that the class of emitter matches that of the
annotation. This allows for correct behavior when mixing different
emitters, e.g., -E high -e middle.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>