Switch to HowToSerialize for Emission (#1405)
* Fix emit{Firrtl,Verilog} for CustomFileEmission

Change ChiselStage helper methods for emitting FIRRTL (emitFirrtl) and
Verilog (emitVerilog) to look for Circuit and Verilog annotations
instead of DeletedAnnotations. This is needed after migrating to the
CustomFileEmission mixin in FIRRTL where FIRRTL will no longer delete
emitter annotations.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Use CustomFileEmission for ChiselCircuitAnnotation

Removes the explicit chisel3.phases.Emitter and instead does emission
with a CustomFileEmission mixin to ChiselCircuitAnnotation. This then
prevents the need for passing around DeletedAnnotations. As a
consequence, I removed an unnecessary run of a second Converter in the
Driver.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Fix tests for use of CustomFileEmission trait

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Fixes for newer CustomFileEmission API

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>