Deprecate support for Scala 2.11 (#1842)
* Deprecate support for Scala 2.11

Printed as warning in transform, can be suppressed via CLI

* Place version deprecation message in object

Refactor CheckScalaVersion to place the deprecation message inside the
companion object. This lets the object be reused in other places, like
in tests.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Check Scala 2.11 deprecation messages in tests

Fix FirrtlMainSpec to check that the Scala 2.11 deprecation message
added by the CheckScalaVersion transform shows up on stdout, but only
if tests are running under 2.11.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

Co-authored-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>