Deprecate CompilerAnnotation (#1870)
* CompilerAnnotation$ emits RunFirrtlTransform

Change the CompilerAnnotation object to emit
RunFirrtlTransformAnnotations containing the associated emitter.

This requires a fix in the Driver compatibility layer to know how to
enable one-file-per module emission if either a CompilerAnnotation or
a RunFirrtlTransformAnnotation(_: Emitter) is present.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Add ConvertCompilerAnnotation phase

Add a phase, ConvertCompilerAnnotation, that converts a
CompilerAnnotation to a RunFirrtlTransformAnnotation. This provides a
warning to the user if this path is taken.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Add test of ConvertCompilerAnnotation

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Deprecate CompilerAnnotation$, move helper methods

Deprecate the CompilerAnnotation companion object and move it's
private utility inside the RunFirrtlTransformAnnotation companion
object.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Make ConvertCompilerAnnotations private[firrtl]

Make this phase private to avoid adding a deprecation warning. Also,
remove an unused string value.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Fix incorrect string in test

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>

* Add test that '-X verilog', no emitter yields file

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>