Fix bug where refs may not get set for Records (#1645)
This requires a combination of things, but it happens to be a
combination used by Diplomacy in Rocket Chip. It must be a Record in
compatibility code with Vecs as fields and a mix of components with and
without set directions.

(cherry picked from commit 0979133b2e67c82f916c2739bbda5df1c692a1d3)