Add Annotation passing support, set firtool CLI
Change ChiselStage to pass annotations to firtool via a file.  Also, add
logic to remove lots of annotations and convert them to equivalent
firtool options.

Any annotations that we don't know what to do with are passed on to firtool to
let it deal with them.  This has the effect of both communicating things like
DontTouchAnnotation and also providing information to the user that they are
passing annotations that firtool doesn't know how to handle.  To do the latter
checking "-warn-on-unprocessed-annotations" is turned on.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>