Add ChiselEnum.safe factory method and avoid warning
Previously, ChiselEnum would warn any time a UInt is converted to an
Enum. There was no way to suppress this warning. Now there is a factory
method (`.safe`) that does not warn and returns (Enum, Bool) where the
Bool is the result of calling .isValid on an Enum object. The regular
UInt cast is also now smarter and will not warn if all bitvectors of the
width of the Enum are legal states.

(cherry picked from commit 5fe539c707c88eedbb112f5c6bcea1dfe1d52169)

# Conflicts:
#	core/src/main/scala/chisel3/StrongEnum.scala
#	src/test/scala/chiselTests/ChiselSpec.scala